code clean up and add padding on screens + header
