+++ src/components/atoms/AppContainer/AppContainer.styles.tsx
@@ -1,0 +1,10 @@
+import {StyleSheet} from 'react-native';
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    padding: 10,
+  },
+});
+
+export default styles;
