+++ src/components/atoms/ButtonWithIcon/index.tsx
@@ -3,7 +3,8 @@ import {
   View as DefaultView,
   Pressable as DefaultPressable,
 } from 'react-native';
-import {Icon as DefaultIcon, IconProps} from '@rneui/themed';
+import {IconProps} from '@rneui/themed';
+import {Icon} from '../Icon';
 
 type ThemeProps = {
   lightColor?: string;
@@ -21,7 +22,7 @@ export const ButtonWithIcon = (props: ButtonWithIconProps) => {
   return (
     <DefaultView style={style}>
       <DefaultPressable onPress={onPress}>
-        <DefaultIcon name={name} type={type} color={color} />
+        <Icon name={name} type={type} color={color} />
       </DefaultPressable>
     </DefaultView>
   );
