+++ src/components/atoms/Divider/index.tsx
@@ -1,0 +1,5 @@
+import {Divider as DefaultDivider, DividerProps} from '@rneui/base';
+
+export const Divider = (props: DividerProps) => {
+  return <DefaultDivider {...props} />;
+};
