+++ src/components/atoms/Icon/index.tsx
@@ -1,0 +1,5 @@
+import {Icon as DefaultIcon, IconProps} from '@rneui/themed';
+
+export const Icon = (props: IconProps) => {
+  return <DefaultIcon {...props} />;
+};
