+++ src/components/atoms/ScreenHeader/ScreenHeader.styles.tsx
@@ -1,0 +1,19 @@
+import {StyleSheet} from 'react-native';
+
+const styles = StyleSheet.create({
+  container: {
+    paddingVertical: 15,
+    paddingHorizontal: 10,
+  },
+  title: {
+    fontSize: 24,
+    marginRight: 4,
+    fontWeight: 'bold',
+  },
+  subtitle: {
+    fontSize: 14,
+    marginRight: 4,
+  },
+});
+
+export default styles;
