+++ src/components/atoms/index.ts
@@ -1,7 +1,10 @@
-export * from './Accordion'
+export * from './Accordion';
 export * from './Button';
 export * from './ButtonWithIcon';
 export * from './DrawerItem';
 export * from './MemberCard';
 export * from './ScreenHeader';
 export * from './TextInput';
+export * from './AppContainer';
+export * from './Icon';
+export * from './Divider';
