+++ src/components/molecules/CommunityDetailsModal/index.tsx
@@ -1,6 +1,7 @@
 import {useState} from 'react';
 import {Modal, Pressable, Text, TouchableHighlight, View} from 'react-native';
-import {Icon} from '@rneui/themed';
+
+import {Icon} from 'components/atoms';
 
 import styles from './CommunityDetailsModal.styles';
