+++ src/components/molecules/DrawerAccordion/index.tsx
@@ -1,9 +1,9 @@
 import {Text, View} from 'react-native';
 import {useState} from 'react';
-import {Icon, ListItem} from '@rneui/base';
+import {ListItem} from '@rneui/base';
 import {DrawerNavigationHelpers} from '@react-navigation/drawer/lib/typescript/src/types';
 
-import {DrawerItem} from 'components/atoms';
+import {DrawerItem, Icon} from 'components/atoms';
 import styles from './DrawerAccordion.styles';
 
 export type DrawerAccordionItem = {
