+++ src/components/molecules/Search/index.tsx
@@ -1,7 +1,8 @@
 import {useState} from 'react';
-import {View, StyleSheet, StyleProp, ViewStyle} from 'react-native';
+import {View, StyleProp, ViewStyle} from 'react-native';
 
 import {ButtonWithIcon, TextInput} from 'components/atoms';
+import styles from './Search.styles';
 
 type SearchProps = {
   onSearch: (searchText: string) => void;
@@ -29,25 +30,3 @@ export const Search = (props: SearchProps) => {
     </View>
   );
 };
-
-const styles = StyleSheet.create({
-  searchContainer: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    margin: 5,
-  },
-  searchInput: {
-    flex: 0.8,
-    borderRadius: 50,
-    borderColor: '#dadce0',
-    borderStyle: 'solid',
-    borderWidth: 1,
-    paddingHorizontal: 20,
-  },
-  searchButton: {
-    flex: 0.15,
-    justifyContent: 'center',
-    borderRadius: 50,
-    backgroundColor: '#6AA2DC',
-  },
-});
