+++ src/components/screens/CEC/CECDashboardScreen/CECDashboardScreen.styles.tsx
@@ -1,0 +1,17 @@
+import {StyleSheet} from 'react-native';
+
+const styles = StyleSheet.create({
+  container: {
+    flexDirection: 'row',
+    alignItems: 'center',
+  },
+  text: {
+    fontSize: 24,
+    marginRight: 4,
+  },
+  tooltipText: {
+    color: 'white',
+  },
+});
+
+export default styles;
