+++ src/components/screens/CEC/CECDashboardScreen/index.tsx
@@ -1,38 +1,12 @@
-import {StyleSheet, Text, View} from 'react-native';
-import {useState} from 'react';
-import {Icon, Tooltip} from '@rneui/base';
+import {Text} from 'react-native';
+
+import {AppContainer} from 'components/atoms';
+import styles from './CECDashboardScreen.styles';
 
 export const CECDashboardScreen = () => {
-  const [open, setOpen] = useState(false);
-
   return (
-    <View style={styles.container}>
+    <AppContainer style={styles.container}>
       <Text style={styles.text}>CEC Requests</Text>
-      <Tooltip
-        skipAndroidStatusBar
-        withOverlay={false}
-        visible={open}
-        onOpen={() => setOpen(true)}
-        onClose={() => setOpen(false)}
-        popover={<Text style={styles.tooltipText}>Insert info here!</Text>}
-      >
-        <Icon name="info" type="material" />
-      </Tooltip>
-    </View>
+    </AppContainer>
   );
 };
-
-const styles = StyleSheet.create({
-  container: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    padding: 5,
-  },
-  text: {
-    fontSize: 24,
-    marginRight: 4,
-  },
-  tooltipText: {
-    color: 'white',
-  },
-});
