+++ src/components/screens/CommunityMembersScreen/index.tsx
@@ -2,7 +2,7 @@ import {useEffect, useState} from 'react';
 import {FlatList, Text, View} from 'react-native';
 import {RouteProp, useRoute} from '@react-navigation/native';
 
-import {MemberCard, ScreenHeader} from 'components/atoms';
+import {AppContainer, MemberCard, ScreenHeader} from 'components/atoms';
 import {Search} from 'components/molecules';
 import {Community, Employee} from 'models/business';
 import {RootNativeStackParamList} from '../../../@types/navigation';
@@ -49,7 +49,7 @@ export const CommunityMembersScreen = () => {
   };
 
   return (
-    <>
+    <AppContainer>
       <ScreenHeader title={name} subtitle={`Managed By: ${managerName}`} />
       {members?.length === 0 ? (
         <NoResult />
@@ -70,6 +70,6 @@ export const CommunityMembersScreen = () => {
           )}
         </>
       )}
-    </>
+    </AppContainer>
   );
 };
