+++ src/components/screens/DashboardScreen/index.tsx
@@ -1,4 +1,4 @@
-import {ScreenHeader} from 'components/atoms';
+import {AppContainer, ScreenHeader} from 'components/atoms';
 import {MemberDetailForm} from 'components/organisms';
 import {useUserDataProvider} from 'providers/UserDataProvider';
 
@@ -7,9 +7,9 @@ export const DashboardScreen = () => {
 
   const dashboardTitle = `Hi, ${user?.username}`;
   return (
-    <>
+    <AppContainer>
       <ScreenHeader title={dashboardTitle} />
       <MemberDetailForm />
-    </>
+    </AppContainer>
   );
 };
