+++ src/components/screens/MembersScreen/MembersScreen.styles.tsx
@@ -1,0 +1,9 @@
+import {StyleSheet} from 'react-native';
+
+const styles = StyleSheet.create({
+  text: {
+    fontSize: 24,
+  },
+});
+
+export default styles;
