+++ src/components/screens/MembersScreen/index.tsx
@@ -1,11 +1,12 @@
-import {StyleSheet, Text} from 'react-native';
+import {Text} from 'react-native';
+
+import {AppContainer} from 'components/atoms';
+import styles from './MembersScreen.styles';
 
 export const Members = () => {
-  return <Text style={styles.text}>Members</Text>;
+  return (
+    <AppContainer>
+      <Text style={styles.text}>Members</Text>
+    </AppContainer>
+  );
 };
-
-const styles = StyleSheet.create({
-  text: {
-    fontSize: 24,
-  },
-});
