+++ src/navigation/MainNavigation.styles.tsx
@@ -6,9 +6,15 @@ const styles = StyleSheet.create({
     height: 60,
     justifyContent: 'center',
     alignItems: 'center',
-    paddingVertical: 5,
+    paddingVertical: 10,
     borderBottomColor: '#ccc',
     borderBottomWidth: 2,
+  },
+  headerRight: {
+    paddingRight: 10,
+  },
+  headerLeft: {
+    paddingLeft: 10,
   },
 });
